pyproject.toml
[project]
name = "shopping-agent"
version = "0.1.0"
description = "AI-powered shopping agent using Google's Universal Commerce Protocol (UCP)"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
]

[project.scripts]
shopping-agent = "shopping_agent.main:main"
shopping-agent-server = "shopping_agent.mock_server:run"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
